+++ grants_tagger_light/evaluation/evaluate_human.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import typer
 from sklearn.metrics import classification_report, f1_score
 
-from grants_tagger_light.evaluation.utils import load_data
+from grants_tagger_light.utils import load_data
 
 
 def evaluate_human(data_path, label_binarizer_path):
